# clean-arch/docs/swagger.yaml
basePath: /
definitions:
  requests.CreateProduct:
    properties:
      name:
        type: string
      price:
        type: number
      user_id:
        type: integer
    type: object
  responses.GetProduct:
    properties:
      id:
        example: 10
        type: integer
      name:
        example: Riksa
        type: string
      price:
        example: 20000
        type: number
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /product:
    post:
      consumes:
      - application/json
      description: create product
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create product
      tags:
      - product
  /products:
    get:
      consumes:
      - application/json
      description: get products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.GetProduct'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get products
      tags:
      - product
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"